Move tests over from scala/scala
All parallel collections tests are now in two subprojects, for JUnit and
ScalaCheck tests respectively. The main codebase is moved to the `core`
subproject. We are currently using a ScalaCheck build for 2.12 but it
will become incompatible once we start getting serious with the new
collections library, so this is not a long-term solution. We may have
to insource ScalaCheck into this project as well, just as we did for
scala/scala. At least the separation of ScalaCheck and JUnit tests will
make it easy to temporarily disable the ScalaCheck tests if no
compatible ScalaCheck build is available.

Moving some `CollectionConverters` methods into a new
`CollectionConvertersLowPriority` trait was necessary to fix a bug
caused by ambiguous implicits that showed up in the tests.